refactor: migrate layouts to src dir
