+++ src/layouts/threecolumn/ThreeColumnLayout.stories.tsx
@@ -1,13 +1,13 @@
-// stories/DashboardLayout.stories.js
+// stories/ThreeColumnLayout.stories.js
 import React from 'react';
-import { DashboardLayout } from '@/components/layouts/dashboard/DashboardLayout';
-import { showcaseParameters } from '../../../../stories/examples/showcaseParameters';
+import { ThreeColumnLayout } from 'src/layouts/threecolumn/ThreeColumnLayout';
+import { showcaseParameters } from '../../../stories/examples/showcaseParameters';
 import { Box } from '@mui/system';
 import { NextPageWithLayout } from '@/types/next-page';
 
 export default {
   title: 'Showcase/Layouts',
-  component: DashboardLayout,
+  component: ThreeColumnLayout,
   parameters: showcaseParameters,
 };
 
@@ -23,14 +23,16 @@ const dummybox = {
   display: 'flex',
 };
 
-export const Dashboard = () => {
+export const ThreeColumns = () => {
   const TestPage: NextPageWithLayout = () => (
     <Box sx={dummybox}>Main content goes here</Box>
   );
   TestPage.getLayout = (page) => page; // Add a default getLayout function to your page
 
-  const LayoutTestPage = DashboardLayout(TestPage, {
-    title: 'DashboardLayout Example',
+  const LayoutTestPage = ThreeColumnLayout(TestPage, {
+    title: 'ThreeColumnLayout Example',
+    leftSidebar: <Box sx={dummybox}>Left Sidebar Content</Box>,
+    rightSidebar: <Box sx={dummybox}>Right Sidebar Content</Box>,
   });
 
   // Simulate a Next.js page by calling the getLayout function
